+++ backend/Cargo.toml
@@ -1,14 +1,20 @@
 [package]
-name = "backend"
+name = "backend-axum"
 version = "0.1.0"
 edition = "2021"
 
 [dependencies]
-rocket = { version = "=0.5.0-rc.3", features = ["json"] }
-serde = { version = "1.0.130", features = ["derive"] }
-serde_json = "1.0.68"
-rocket_db_pools = { version = "=0.1.0-rc.3", features = ["sqlx_postgres"]}
+axum = "0.6.18"
+tokio = { version = "1.28.2", features = ["full"] }
 dotenvy = "0.15.7"
-env_logger = "0.9"
+
+serde = { version = "1.0.164", features = ["derive"] }
+serde_json = "1.0.68"
+
 sqlx = { version = "0.6.3", features = ["postgres", "runtime-tokio-rustls", "migrate", "uuid", "chrono", "offline"]}
 chrono = { version = "0.4.26", features = ["serde"] }
+tower = {version = "0.4.12", features = ["full"] }
+tower-http = {version = "0.4.0", features = ["full"] }
+hyper = { version = "0.14.13", features = ["full"] }
+tracing = "0.1"
+tracing-subscriber = { version = "0.3", features = ["fmt", "env-filter"] }
