--- backend/Rocket.toml
@@ -1,10 +1,0 @@
-[debug]
-address = "127.0.0.1"
-port = 8000
-keep_alive = 5
-limits = { forms = 32768 }
-
-[release]
-address = "0.0.0.0"
-port = 8000
-limits = { forms = 32768 }
