--- backend/static/404.html
@@ -1,18 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <title>Page Not Found</title>
-    <style>
-      body {
-        text-align: center;
-        padding: 50px;
-      }
-    </style>
-  </head>
-  <body>
-    <h1>Oops!</h1>
-    <p>Sorry, we can't find the page you're looking for.</p>
-    <a href="/">Go back to home</a>
-  </body>
-</html>
